move shared jest.config out of hierarchy reduces tests by half (which is correct) - tests pass
