+++ packages/apollo-boost/jest.config.js
@@ -1,4 +1,4 @@
-const pkg = require('../jest.config.settings');
+const pkg = require('../../config/jest.config.settings');
 
 module.exports = {
   ...pkg,
